+++ ergo-lib/src/ast/calc_blake2b256.rs
@@ -14,7 +14,7 @@ use crate::util::AsVecU8;
 use super::expr::Expr;
 use super::expr::InvalidArgumentError;
 use super::value::CollKind;
-use super::value::CollPrim;
+use super::value::NativeColl;
 use super::value::Value;
 
 #[derive(PartialEq, Eq, Debug, Clone)]
@@ -44,7 +44,7 @@ impl Evaluable for CalcBlake2b256 {
         let input_v = self.input.eval(env, ctx)?;
         match input_v.clone() {
             Value::Coll(coll) => match coll {
-                CollKind::Primitive(CollPrim::CollByte(coll_byte)) => {
+                CollKind::NativeColl(NativeColl::CollByte(coll_byte)) => {
                     let expected_hash: Vec<u8> =
                         blake2b256_hash(coll_byte.as_vec_u8().as_slice()).0.to_vec();
                     Ok(expected_hash.into())