+++ ergo-lib/src/eval/expr.rs
@@ -32,6 +32,8 @@ impl Evaluable for Expr {
             Expr::And(op) => op.eval(env, ctx),
             Expr::Or(op) => op.eval(env, ctx),
             Expr::LogicalNot(op) => op.eval(env, ctx),
+            Expr::Map(op) => op.eval(env, ctx),
+            Expr::Filter(op) => op.eval(env, ctx),
         }
     }
 }